requirements.txt
openai
rich
python-dotenv
streamlit
gTTS
